perf: 调整model包文件结构
