+++ pkg/model/asset_test.go
